requirements.txt
aniso8601==9.0.0
argon2-cffi==20.1.0
click==7.1.2
et-xmlfile==1.0.1
Flask==1.1.2
Flask-Cors==3.0.10
Flask-JWT==0.3.2
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.4
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.3
MarkupSafe==1.1.1
marshmallow==3.10.0
mysql-connector==2.2.9
numpy==1.20.1
pandas==1.2.2
passlib==1.7.4
PyJWT==1.4.2
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
SQLAlchemy==1.3.23
Werkzeug==1.0.1
xlrd==2.0.1
newrelic
uwsgi